refactored code face verification, stream result
