+++ src/server/verification/verificationAPI.js
@@ -3,7 +3,6 @@ import { Router } from 'express'
 import passport from 'passport'
 import _ from 'lodash'
 import moment from 'moment'
-import multer from 'multer'
 import type { LoggedUser, StorageAPI, UserRecord, VerificationAPI } from '../../imports/types'
 import AdminWallet from '../blockchain/AdminWallet'
 import { onlyInEnv, wrapAsync } from '../utils/helpers'
@@ -13,16 +12,12 @@ import { sendOTP, generateOTP } from '../../imports/otp'
 import conf from '../server.config'
 import { GunDBPublic } from '../gun/gun-middleware'
 import { Mautic } from '../mautic/mauticAPI'
-import fs from 'fs'
 import W3Helper from '../utils/W3Helper'
 import gdToWei from '../utils/gdToWei'
 import txManager from '../utils/tx-manager'
-import humanAPI from './faceRecognition/human'
+import Verifications from './verification'
 
-const fsPromises = fs.promises
 const setup = (app: Router, verifier: VerificationAPI, storage: StorageAPI) => {
-  var upload = multer({ dest: 'uploads/' }) // to handle blob parameters of faceReco
-
   /**
    * @api {post} /verify/facerecognition Verify users face
    * @apiName Face Recognition
@@ -43,60 +38,23 @@ const setup = (app: Router, verifier: VerificationAPI, storage: StorageAPI) => {
     wrapAsync(async (req, res, next) => {
       const { sessionId, images } = req.body
       const { user } = req
-      const userId = user.identifier
       const log = req.log
-
       let result = { ok: 0 }
 
-      if (userId === undefined || sessionId === undefined || images === undefined) {
+      if (user.identifier === undefined || sessionId === undefined || images === undefined) {
         return res.status(400).send({ ok: 0, error: 'invalid input', isVerified: false })
       }
-
       GunDBPublic.gun
         .get(sessionId)
         .get('isStarted')
-        .put(true) // publish initialized data to subscribers
+        .put(true)
       log.debug('written FR status to gun', { data: await GunDBPublic.gun.get(sessionId) })
+
+      const imagesBase64 = images.map(img => img.base64)
 
       if (!user.isVerified && !conf.skipFaceRecognition) {
         try {
-          const enrollHandler = async (userId, sessionId, data) => {
-            log.debug('enrollHandler', data)
-            if (data.ok && data.result) {
-              log.debug('Whitelisting new user', user)
-              try {
-                await Promise.all([
-                  AdminWallet.whitelistUser(user.gdAddress, user.profilePublickey),
-                  storage
-                    .updateUser({ identifier: user.loggedInAs, isVerified: true })
-                    .then(updatedUser => log.debug('updatedUser:', updatedUser))
-                ])
-                await GunDBPublic.gun
-                  .get(sessionId)
-                  .get('isWhitelisted')
-                  .put(true) // publish to subscribers
-              } catch (e) {
-                log.error('Whitelisting failed', e)
-                GunDBPublic.gun.get(sessionId).put({ isWhitelisted: false, isError: e.message })
-              }
-
-              result = {
-                ok: 1,
-                isVerified: true,
-                enrollResult: data.isEnroll
-              }
-            } else {
-              log.error('Facerecognition error:', data.error)
-              GunDBPublic.gun.get(sessionId).put({ isNotDuplicate: false, isLive: false, isError: data.error })
-              result = { ok: 1, error: data.error, isVerified: false }
-            }
-          }
-
-          const imagesBase64 = images.map(img => {
-            return img.base64
-          })
-
-          await humanAPI.addIfUniqueAndAlive(userId, sessionId, imagesBase64, enrollHandler)
+          result = await Verifications.verifyUser(user, sessionId, imagesBase64, storage)
         } catch (e) {
           log.error('Facerecognition error:', { e })
           GunDBPublic.gun.get(sessionId).put({ isNotDuplicate: false, isLive: false, isError: e.message })
@@ -111,6 +69,7 @@ const setup = (app: Router, verifier: VerificationAPI, storage: StorageAPI) => {
           enrollResult: { alreadyEnrolled: true, enrollmentIdentifier: true }
         }
       }
+
       res.json(result)
     })
   )
